requirements.txt
# ./requirements.txt
rich>=10.0.0
PyYAML>=6.0
watchdog>=2.1.0